requirements.txt
# --- core ---
streamlit>=1.37,<2
python-dotenv>=1.0,<2
pydantic>=2.5,<3

# --- LLM / LangChain ---
openai>=1.40,<2
langchain>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-openai>=0.2,<0.4
langchain-chroma>=0.2,<0.4
langchain-text-splitters>=0.3,<0.4
tiktoken>=0.7,<1

# --- Vector store ---
chromadb>=1.0,<2
duckdb>=1.0,<2

# --- Loaders / parsing ---
pdfplumber>=0.11,<0.12
pypdf>=4,<7
Pillow>=10,<12
beautifulsoup4>=4.12,<5
requests>=2.31,<3

# --- (dev) tests ---
pytest>=8,<9